first pass at simplifying is a mess
