+++ src/components/Course.js
@@ -1,10 +1,23 @@
-import React, { useState, useEffect } from "react";
-import { Switch, Route, useRouteMatch, useParams } from "react-router-dom";
-import { makeStyles } from "@material-ui/core/styles";
+import React, { useState, useEffect, useRef } from "react";
+import {
+  Switch,
+  Route,
+  Link as RouterLink,
+  useRouteMatch,
+  useParams,
+} from "react-router-dom";
+import { makeStyles, useTheme } from "@material-ui/core/styles";
+import useMediaQuery from "@material-ui/core/useMediaQuery";
 import Container from "@material-ui/core/Container";
+import Box from "@material-ui/core/Box";
+import Zoom from "@material-ui/core/Zoom";
+import Fab from "@material-ui/core/Fab";
+import ArrowBackIcon from "@material-ui/icons/ArrowBack";
+import ArrowForwardIcon from "@material-ui/icons/ArrowForward";
 import CourseHeader from "./CourseHeader";
-import ItemList from "./ItemList";
+import Item from "./Item";
 import Score from "./Score";
+import Review from "./Review";
 import FeedbackModal from "./FeedbackModal";
 import NotFound from "./NotFound";
 import isAnswerCorrect from "../utils/isAnswerCorrect";
@@ -19,11 +32,37 @@ const useStyles = makeStyles((theme) => ({
   container: {
     margin: theme.spacing(0, "auto", 12, "auto"),
   },
+  item: {
+    marginTop: theme.spacing(3),
+  },
+  fabRight: {
+    margin: 0,
+    position: "fixed",
+    top: "auto",
+    right: theme.spacing(3),
+    bottom: theme.spacing(3),
+    left: "auto",
+  },
+  fabRightIcon: {
+    marginLeft: theme.spacing(1),
+  },
+  fabLeft: {
+    margin: 0,
+    position: "fixed",
+    top: "auto",
+    left: theme.spacing(3),
+    bottom: theme.spacing(3),
+    right: "auto",
+  },
+  fabLeftIcon: {
+    marginRight: theme.spacing(1),
+  },
 }));
 
 export default function Course() {
   const classes = useStyles();
-  let { path } = useRouteMatch();
+  const theme = useTheme();
+  let { path, url } = useRouteMatch();
   let { courseId } = useParams();
   const [course, setCourse] = useState(null);
   const [userEmail, setUserEmail] = useState(""); // to ID user
@@ -31,8 +70,26 @@ export default function Course() {
   const [showSolutions, setShowSolutions] = useState(false);
   const [showFeedbackModal, setShowFeedbackModal] = useState(false);
   const [show404, setShow404] = useState(false);
-  const [orientation, setOrientation] = useState("horizontal");
   const [itemNumber, setItemNumber] = useState(0);
+  const notOnMobile = useMediaQuery(theme.breakpoints.up("sm"));
+  const continueRef = useRef(null);
+
+  useEffect(() => {
+    // if the ref exists, create keyboard shortcut
+    if (continueRef && continueRef.current) {
+      const listener = (event) => {
+        if (event.code === "Enter" || event.code === "NumpadEnter") {
+          event.preventDefault();
+          continueRef.current.click();
+        }
+      };
+      document.addEventListener("keydown", listener);
+
+      return () => {
+        document.removeEventListener("keydown", listener);
+      };
+    }
+  });
 
   useEffect(() => {
     getCourseFromFirestore(courseId)
@@ -74,12 +131,6 @@ export default function Course() {
     }
   }, [courseId, userEmail, answers, course, showSolutions]);
 
-  // if answers are shown, switch to vertical view mode
-  // and set progress percentage to 100
-  useEffect(() => {
-    if (showSolutions) setOrientation("vertical");
-  }, [showSolutions]);
-
   const getSolution = (itemId) => {
     const item = course.items.filter((i) => i.id === itemId)[0];
     return item.solution;
@@ -104,8 +155,22 @@ export default function Course() {
     ]);
   };
 
+  const handleFinishCourse = () => {
+    setItemNumber(course.items.length);
+    // use timeout so progress bar has time to advance to 100%
+    setTimeout(() => {
+      setShowSolutions(true);
+    }, 500);
+  };
+
+  // if the course isn't loaded yet, return null
+  if (!course) return null;
+
   // if the course isn't found, show 404
   if (show404) return <NotFound type="course" />;
+
+  // single in on the current item
+  const item = course.items[itemNumber];
 
   // compute progress percentage
   const progress = Math.min(
@@ -114,50 +179,135 @@ export default function Course() {
   );
 
   return (
-    course && (
-      <>
-        <CourseHeader
-          courseTitle={course.title}
-          progress={progress}
-          orientation={orientation}
-          setOrientation={setOrientation}
-          setShowFeedbackModal={setShowFeedbackModal}
-        />
-        <FeedbackModal
-          open={showFeedbackModal}
-          setOpen={setShowFeedbackModal}
-          sentFrom="course"
-          answers={answers}
-        />
-        <Container className={classes.container}>
-          <Switch>
-            <Route exact path={path}>
-              <ItemList
-                items={course.items}
-                answers={answers}
-                onChangeAnswer={handleChangeAnswer}
-                showSolutions={showSolutions}
-                setShowSolutions={setShowSolutions}
-                orientation={orientation}
-                itemNumber={itemNumber}
-                setItemNumber={setItemNumber}
-                userEmail={userEmail}
-                setUserEmail={setUserEmail}
-              />
-            </Route>
-            <Route exact path={`${path}/score`}>
-              <Score
-                message={course.finalMessage}
-                finalCta={course.finalCta}
-                answers={answers}
-              />
-            </Route>
-            <Route path={`${path}/*`}>
-              <NotFound type="page" />
-            </Route>
-          </Switch>
-        </Container>
-      </>
-    )
+    <>
+      <CourseHeader
+        courseTitle={course.title}
+        progress={progress}
+        showProgress={!showSolutions}
+        setShowFeedbackModal={setShowFeedbackModal}
+      />
+      <Container className={classes.container}>
+        <Switch>
+          <Route exact path={path}>
+            {showSolutions ? (
+              <>
+                <Review
+                  items={course.items}
+                  answers={answers}
+                  userEmail={userEmail}
+                />
+                <Zoom in>
+                  <Fab
+                    className={classes.fabRight}
+                    component={RouterLink}
+                    to={`${url}/score`}
+                    variant="extended"
+                    color="primary"
+                    aria-label="submit"
+                  >
+                    Back to my score
+                    <ArrowForwardIcon className={classes.fabRightIcon} />
+                  </Fab>
+                </Zoom>
+              </>
+            ) : (
+              <>
+                {item && (
+                  <>
+                    <Box className={classes.item}>
+                      <Item
+                        key={item.id}
+                        item={item}
+                        answer={
+                          answers &&
+                          answers.filter((a) => a.itemId === item.id)[0]
+                        }
+                        onChangeAnswer={handleChangeAnswer}
+                        userEmail={userEmail}
+                        setUserEmail={setUserEmail}
+                      />
+                    </Box>
+                    {itemNumber > 0 && (
+                      <Zoom in>
+                        <Fab
+                          className={classes.fabLeft}
+                          onClick={() => setItemNumber(itemNumber - 1)}
+                          variant={notOnMobile ? "extended" : "round"}
+                          color="primary"
+                          aria-label="go back"
+                        >
+                          {notOnMobile ? (
+                            <>
+                              <ArrowBackIcon className={classes.fabLeftIcon} />{" "}
+                              Go back
+                            </>
+                          ) : (
+                            <ArrowBackIcon />
+                          )}
+                        </Fab>
+                      </Zoom>
+                    )}
+                    <Zoom in>
+                      {itemNumber < course.items.length - 1 ? (
+                        <Fab
+                          ref={continueRef}
+                          className={classes.fabRight}
+                          onClick={() => setItemNumber(itemNumber + 1)}
+                          variant={notOnMobile ? "extended" : "round"}
+                          disabled={item.type === "Email" && !userEmail}
+                          color="primary"
+                          aria-label="continue"
+                        >
+                          {notOnMobile ? (
+                            <>
+                              Continue{" "}
+                              <ArrowForwardIcon
+                                className={classes.fabRightIcon}
+                              />
+                            </>
+                          ) : (
+                            <ArrowForwardIcon />
+                          )}
+                        </Fab>
+                      ) : (
+                        <Fab
+                          ref={continueRef}
+                          className={classes.fabRight}
+                          component={RouterLink}
+                          to={`${url}/score`}
+                          onClick={handleFinishCourse}
+                          variant="extended"
+                          color="primary"
+                          aria-label="submit"
+                        >
+                          I'm all done!
+                          <ArrowForwardIcon className={classes.fabRightIcon} />
+                        </Fab>
+                      )}
+                    </Zoom>
+                  </>
+                )}
+              </>
+            )}
+          </Route>
+          <Route exact path={`${path}/score`}>
+            <Score
+              message={course.finalMessage}
+              finalCta={course.finalCta}
+              answers={answers}
+            />
+          </Route>
+          <Route path={`${path}/*`}>
+            <NotFound type="page" />
+          </Route>
+        </Switch>
+      </Container>
+      <FeedbackModal
+        open={showFeedbackModal}
+        setOpen={setShowFeedbackModal}
+        sentFrom="course"
+        answers={answers}
+      />
+    </>
   );
 }
